chrefactor: extract elements
