+++ src/components/element-x-axis-tooltip.ts
@@ -1,0 +1,68 @@
+import { Colors } from "../lib/vega-colours";
+import { Element } from "../types/element";
+import { format } from "date-fns";
+
+export function addXAxisPath(
+  ctx: CanvasRenderingContext2D,
+  xScale: any,
+  yScale: any,
+  position: [number, number] | null
+) {
+  if (position) {
+    const height = 24.5;
+    const x = position[0];
+
+    ctx.font = `12px monospace`;
+    ctx.textAlign = "center";
+    ctx.textBaseline = "middle";
+
+    const value = xScale.invert(x);
+    const xPad = 5;
+    const text = format(value, "HH:mm");
+    const textWidth = ctx.measureText(text).width;
+    const rectWidth = textWidth + xPad * 2;
+    const rectHeight = 19;
+
+    let xAdjusted = x;
+
+    if (x - rectWidth / 2 < 0) {
+      xAdjusted = rectWidth / 2;
+    }
+
+    if (x + rectWidth / 2 > xScale.range()[1]) {
+      xAdjusted = xScale.range()[1] - rectWidth / 2;
+    }
+
+    ctx.beginPath();
+    ctx.moveTo(xAdjusted, height - rectHeight - 5);
+    ctx.lineTo(xAdjusted + 5, height - rectHeight);
+    ctx.lineTo(xAdjusted + rectWidth / 2, height - rectHeight);
+    ctx.lineTo(xAdjusted + rectWidth / 2, height);
+    ctx.lineTo(xAdjusted - rectWidth / 2, height);
+    ctx.lineTo(xAdjusted - rectWidth / 2, height - rectHeight);
+    ctx.lineTo(xAdjusted - 5, height - rectHeight);
+    ctx.closePath();
+
+    ctx.fillStyle = Colors.GRAY_DARK_1;
+    ctx.strokeStyle = Colors.GRAY_LIGHT_1;
+    ctx.fill();
+    ctx.stroke();
+    ctx.closePath();
+
+    ctx.beginPath();
+    ctx.fillStyle = Colors.WHITE;
+    ctx.fillText(text, xAdjusted, height - rectHeight / 2);
+    ctx.closePath();
+  }
+}
+
+export class XAxisTooltipElement implements Element {
+  draw(
+    ctx: CanvasRenderingContext2D,
+    xScale: any,
+    yScale: any,
+    position: [number, number] | null
+  ) {
+    addXAxisPath(ctx, xScale, yScale, position);
+  }
+}
