+++ src/types/element.ts
@@ -1,0 +1,8 @@
+export interface Element {
+  draw(
+    ctx: CanvasRenderingContext2D,
+    xScale: any,
+    yScale: any,
+    ...rest: any[]
+  ): void;
+}
